Fix: The check for item persistence was broken.
The code formerly tried to verify that a successfully sent item was
permanently stuck in the queue after disconnect (and not merely a
concurrency illusion) by checking the Weak reference in a nested fashion
like this:

if let Some(_) = previous_weak.upgrade() {
    thread::sleep(Duration::from_secs(1));
    if let Some(_) = previous_weak.upgrade() {
        // ...
    }
}

This is incorrect, as the first upgrade() will increase the strong count
and the nested upgrade() will always succeed.  (The temporary Arc was
not immediately dropped when assigned to Some(_) as I originally assumed
it would be.)

The fix is to check with previous_weak.upgrade().is_some().

Additionally, instead of waiting one second, we now spin up to 100ms
while waiting for the successfully sent item to be dropped from the
disconnected queue.  On my machine, it usually takes ~2us for items to
appear as dropped to the main thread, and any item that still appears to
exist 100ms after disconnect is stuck forever.

Obviously this is all very messy, but it should be good enough for
accurately demonstrating the MPSC race condition.